Fix plot
