+++ docs/img/WaRP-Categories.png
